requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.10.1
autopep8==1.5.7
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.6
cryptography==35.0.0
ecdsa==0.17.0
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
pefile==2021.9.3
pycodestyle==2.7.0
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyOpenSSL==21.0.0
six==1.16.0
toml==0.10.2
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
